+++ src/commands/project.js
@@ -3,7 +3,7 @@ import * as inquirer from '@inquirer/prompts';
 import projectService from '../services/projectService.js';
 import clientService from '../services/clientService.js';
 import clientModel from '../models/client.js';
-import {isProjectInvalid} from '../../validation.js';
+import {isProjectInvalid, isConfirmationInvalid} from '../../validation.js';
 
 //TODO: validate
 
@@ -15,10 +15,7 @@ const createProject = async () => {
   const name = await inquirer.input({
     message: 'What is the name of the project?',
     validate: input => {
-      if (input.toLowerCase() === 'exit') {
-        console.log('Exiting the process...');
-        process.exit(); // TODO: add prpper handling when user what to stop action, whole app
-      }
+      
       let resp;
       if ((resp = isProjectInvalid(input))) return resp;
 
@@ -112,10 +109,28 @@ const deleteProject = async () => {
     },
   });
   if (name === 'Stop') return;
+  const confirmation = await inquirer.input({
+    default: 'no',
+    message: `Are you sure to delete project: ${name} ? (yes/no)`,
+    validate: input => {
+      if (input.toLowerCase() === 'exit') {
+        console.log('Exiting the process...');
+        process.exit(); // TODO: add prpper handling when user what to stop action, whole app
+      }
+      let resp;
+      if ((resp = isConfirmationInvalid(input.toLowerCase()))) return resp;
 
-  const project = data.find(c => c.name === name);
+      return true;
 
-  await projectService.delete(project);
+    },
+  });
+  
+  if(confirmation.toLowerCase()==='yes' || confirmation.toLowerCase()==='y')
+  {
+    const project = data.find(c => c.name === name);
+    await projectService.delete(project);
+      
+  }
 };
 
 export {createProject, selectAllProjects, editProject, deleteProject};
